+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed issue with pinging client when `client_age` < 60
 - Fixed issue with key when client created using different key size
 - Fixed compression flag
+- Fixed issue with token retrieving using default access code
 
 ## [1.0.0-beta.9]
 ### Fixed